Finished my Movable_gate object & changed some FPS files
I had to change 2 files in the FPS game, PlayerCharacter Controller, added a pause option so the player can't move the camera around while paused.
And I did the same thing for WorldSpaceHealthBar because it was throwing errors while the PlayerCharcterController was paused.

I tried changing it in a way where I wouldn't have to change the shared code, but this basically resulted into me rewriting and renaming the entire game, which seemed quite overkill for something so little

Oh and I added an Enemy tag for the Enemies it was part of my solution.